LayoutTests/workers/wasm-long-compile-expected.txt

CONSOLE MESSAGE: Good result
CONSOLE MESSAGE: Finished test
PASS successfullyParsed is true

TEST COMPLETE
